Add centos7 test

+++ test/2-simple-dns01-nslookup.bats
@@ -8,12 +8,16 @@ load '/getssl/test/test_helper.bash'
 # This is run for every test
 setup() {
     export CURL_CA_BUNDLE=/root/pebble-ca-bundle.crt
-    mv /usr/bin/dig /usr/bin/dig.getssl.bak
+    if [ -f /usr/bin/dig ]; then
+        mv /usr/bin/dig /usr/bin/dig.getssl.bak
+    fi
 }
 
 
 teardown() {
-    mv /usr/bin/dig.getssl.bak /usr/bin/dig
+    if [ -f /usr/bin/dig.getssl.bak ]; then
+        mv /usr/bin/dig.getssl.bak /usr/bin/dig
+    fi
 }
